feature(fragment): ensure we export `Fragment` and `InferFragmentType` types
