+++ packages/groq-builder/src/types/public-types.ts
@@ -1,5 +1,7 @@
 import { GroqBuilder } from "../groq-builder";
 import { ResultItem } from "./result-types";
+import { Simplify, Tagged } from "./utils";
+import { ExtractProjectionResult } from "../commands/projection-types";
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
@@ -53,3 +55,13 @@ export type InferResultType<TGroqBuilder extends GroqBuilder> =
 export type InferResultItem<TGroqBuilder extends GroqBuilder> = ResultItem<
   InferResultType<TGroqBuilder>
 >;
+
+export type Fragment<
+  TProjectionMap,
+  TFragmentInput // This is used to capture the type, to be extracted by `InferFragmentType`
+> = Tagged<TProjectionMap, TFragmentInput>;
+
+export type InferFragmentType<TFragment extends Fragment<any, any>> =
+  TFragment extends Fragment<infer TProjectionMap, infer TFragmentInput>
+    ? Simplify<ExtractProjectionResult<TFragmentInput, TProjectionMap>>
+    : never;
